Add bbPress settings to the Genesis Theme Settings customizer panel
Genesis no longer uses a settings page, but has switched over to the customizer for its theme settings.

The bbPress settings were added to the standard settings page in Genesis, but since that is now gone, settings are unavailable for users.

This plugin should, in addition to adding the controls to the theme settings page for older versions of Genesis, add controls to the theme settings customizer panel.

This also tags 1.2.0 and addes change to the changelog in `readme.txt`.